Surface runtime capability summary in keeper editor
