+++ dashboard/src/components/keeper-runtime-model-editor.test.ts
@@ -83,6 +83,64 @@ function makeRuntimeProvider(runtimeId: string, providerName: string, modelName:
     tools_support: true,
     thinking_support: true,
     streaming: true,
+    capabilities_declared: true,
+    max_output_tokens: 8192,
+    supports_tool_choice: true,
+    supports_required_tool_choice: true,
+    supports_named_tool_choice: true,
+    supports_parallel_tool_calls: true,
+    supports_extended_thinking: true,
+    supports_multimodal_inputs: true,
+    supports_image_input: true,
+    supports_audio_input: false,
+    supports_video_input: false,
+    supports_reasoning_budget: true,
+    supports_response_format_json: true,
+    supports_structured_output: true,
+    supports_native_streaming: true,
+    supports_system_prompt: true,
+    supports_caching: true,
+    supports_prompt_caching: true,
+    prompt_cache_alignment: 1024,
+    supports_top_k: true,
+    supports_min_p: true,
+    supports_seed: true,
+    supports_seed_with_images: true,
+    emits_usage_tokens: true,
+    supports_computer_use: false,
+    supports_code_execution: true,
+    source: 'runtime.toml',
+    parameter_policy: {
+      reasoning_toggle_wire: 'responses.reasoning',
+      reasoning_replay_policy: 'preserve',
+      requires_reasoning_replay_on_tool_call: true,
+      ignored_sampling_params: ['top_k'],
+      always_ignored_sampling_params: ['min_p'],
+    },
+    request_config: {
+      source: 'runtime.toml',
+      provider_kind: 'cloud',
+      request_path: '/v1/responses',
+      request_path_targets_responses_api: true,
+      max_tokens: 8192,
+      max_context: 128000,
+      temperature: 0.2,
+      top_p: 0.9,
+      top_k: 40,
+      min_p: 0.05,
+      has_system_prompt: true,
+      enable_thinking: true,
+      preserve_thinking: true,
+      thinking_budget: 4096,
+      resolved_reasoning_effort: 'high',
+      tool_stream: true,
+      tool_choice: { kind: 'required', name: 'inspect' },
+      response_format: { kind: 'json_schema', has_schema: true },
+      has_output_schema: true,
+      cache_system_prompt: true,
+      seed: 7,
+      connect_timeout_s: 30,
+    },
     models: [modelName],
     endpoint_url: 'https://runtime.example/v1',
   }
@@ -150,6 +208,13 @@ describe('KeeperRuntimeModelEditor', () => {
     await flush()
     expect(container.textContent).toContain('Provider A')
     expect(container.textContent).toContain('claude')
+    expect(container.textContent).toContain('caps:declared')
+    expect(container.textContent).toContain('format:json,schema')
+    expect(container.textContent).toContain('prompt-cache@1024')
+    expect(container.textContent).toContain('wire:responses.reasoning')
+    expect(container.textContent).toContain('tool-call-replay:required')
+    expect(container.textContent).toContain('responses-api')
+    expect(container.textContent).toContain('budget:4096')
   })
 
   it('patches runtime_id with the selected runtime and updates shared config', async () => {
@@ -197,6 +262,8 @@ describe('KeeperRuntimeModelEditor', () => {
     // Hint names the runtime assignment surface so the operator can unlock editing.
     expect(container.textContent).toContain('runtime.toml')
     expect(container.textContent).toContain('[runtime.assignments]')
+    expect(container.textContent).toContain('caps:declared')
+    expect(container.textContent).toContain('responses-api')
   })
 
   it('clears a pending selection when the viewed keeper changes', async () => {
